requirements.txt
fastapi==0.99.1 
mangum==0.17.0 
PyMySQL==1.1.1
python-dotenv==0.21.0
boto3==1.26.90